requirements.txt
streamlit
python-dotenv
langchain-google-genai
google-generativeai
easyocr
pdf2image
Pillow
opencv-python-headless
pyperclip
